requirements.txt
Flask
pandas
numpy
yfinance
tensorflow
scikit-learn
xgboost
requests
openai
google-generativeai
python-dotenv
